Workaround for issue #5 (wel operating on flow rate)
Issue #5 is due to the fact that upon initialization well may not product/inject if its pressure is not consistent with the surrounding pressure.
The correct fix would consist in computing first the well state for well operating on flow rate.
This is only a workaround consisting in setting the well head pressure to :
- the minimum local pressure - 1 for production wells
- the maximum local pressure + 1 for production wells
This is available through the python ComPASS.timeloops.check_well_pressure routine.