BugFix: Cherry-pick commit '197844f'
Solve the memory overflow of the capillary pressure in src/core/Flux.F90
and src/core/Jacobian.F90 in the multiphase flow case.
Compute the capillary pressure for both the present phases and the
absent phases in LoisThermoHydro_PressionCapillaire_cv in
src/core/LoisThermoHydro.F90.
Prevent the alignment of the equations for the unknowns located on the
Dirichlet-Dirichlet type nodes in Jacobian_Alignment_man in
src/core/Jacobian.F90.

Need to rewrite the context selection in LoisThermoHydro_Temperature_cv
in src/core/LoisThermoHydro.F90.
Need to prevent the alignment for the hybrid nodes Dirichlet-Neumann
and Neumann-Dirichlet.