PETSc synchronization from python
We initiate a transition towards the use of PETSc from python, relying
on petsc4py. petsc4py is implemented in cython and underlying PETSc objects are accessed
through a handle attribute (cf. Petsc_caster.h).

The use of petsc4py makes that we no longer need to call the PetscFinalize function.

New python classes:
- DistributedSystem: holds various information about the linear system to be solved
(this is independant from petsc).
- Synchronizer: it is the base class to synchronize ghost values, the current
implementation relies on PETSc (simplification of the legacy implementation in Fortran),
but its implementation with mpi4py should be easy (and much simpler)
- We pass a *SolverFramework* object to Newton class which is just a wrapper
around PETSc objects. It must be seen as an attempt to abstract the use of PETSc.
- PetscElements: this class holds PETSc objects used to solve the linear system.
Currently it only holds x the solution vector and wraps the solve method.
- MeshInfo: holds the size of various loci on the mesh (e.g. nodes, fractures...)

TimeLoop wrappers are refactored and wrappers related to SolvePetsc and SyncPetsc modules
are located in specific files.

A few dummy examples are provided in experimental directories.
These are NOT to be considered as self explicit/or documentation pieces.

A lot of global variables are now accessible from the ComPASS main module
(cf. the SimulationInfo structure for example).
This is transitory and they will be progressively encapsulated into classes.