Change subroutines interface in LoisThermoHdyro
An optional flag is introduced to store the result
of the computation of some physical quantities using
either absolute or relative phase ordering.
The default behavior (without flag) remains the previous
behavior. The flag is introduced because multi segmented
wells use absolute phase ordering. Cf. discussion in #537.

A new LoisThermoHydro_viscosite_cv subroutine is introduced
and computes what it says i.e. the viscosity.
The old subroutine, which computed the inverse of the
viscosity, is renamed LoisThermoHydro_UnSurViscosite_cv.