Introduce Python pointer for the dynamic viscosity
I add the implementation of Xalpha to define physical_property(X, f, dXdf).
Physical_property with or without derivatives are distinct.
Dynamic visco is now outside of fluid_properties (linear_water structure) for symmetry.
We must be able to use CompiledPhaseProperty without load_eos (and outside of ComPASS).
Access the python function when the dynamic viscosity is called in a script.
The C wrapper is kept to test the C calls.