Use Python pointer for molar density
It was necessary to remove some "pure" subroutines due to the call of the cpp function.
Modification of the wrappers to call the python function from script, keep cpp wrappers to test the molar_density pointer implementation.
Modification of the tests wich use fluid_properties of the linear_water eos.