Use Python pointer for the volumetric mass density
The volumetric mass density relies on the molar density function and the
components molar masses (which are also possible to modify).
MCP is removed from the volumetric mass density definition, it could
introduce error, refer to issue #552.
Modify the wrappers to call the python function from script, keep the cpp wrappers to test the volumetric_mass_density pointer implementation.