Refactor pipeline management using workflow
Pipelines will always be triggered when scheduled,
triggered manually using the web API or in the context of a merge request.
Pipelines will not be triggered by commits on the default branch
(nightly scheduled pipelines will run CI on the default branch).
Pipelines will not be triggered by commits on branches with name starting with NOCI
(except in the abovementioned cases).